Role based ticket system access Notification system implementation Ticket reminder system User page with roles and connection date Chat system between users Tasks system inside the group projects
The implementation of the notification system was realised using socket.io to send notification in real time. The backend microservice for the notification system was written with express.js and node.js.
The implementation of the chat system was written with express.js and node.js as well and is connected to the notification system.
The tasks system was written in golang as well as the projects microservice
The logger microservice was written with express.js and only admins have access to it.
Now the user can make changes to the tickets they have createed. Both the ticket handler and the admins have access to the ticket system.